# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lucid API Reference",
    "contact": {
      "name": "Roost team",
      "url": "https://github.com/roostmoe/lucid",
      "email": "hello@roost.moe"
    },
    "license": {
      "name": "Apache 2.0 License",
      "url": "https://github.com/roostmoe/lucid/blob/main/LICENSE",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/activation-keys": {
      "get": {
        "tags": [
          "activation-keys"
        ],
        "operationId": "list_activation_keys",
        "responses": {
          "200": {
            "description": "List of activation keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedList_ActivationKey"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "activation-keys"
        ],
        "operationId": "create_activation_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateActivationKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Activation key created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateActivationKeyResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          }
        }
      }
    },
    "/api/v1/activation-keys/{id}": {
      "get": {
        "tags": [
          "activation-keys"
        ],
        "operationId": "get_activation_key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activation key details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivationKey"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "activation-keys"
        ],
        "operationId": "delete_activation_key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Activation key deleted"
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          }
        }
      }
    },
    "/api/v1/agents/register": {
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "POST /api/v1/agents/register",
        "description": "Register a new agent using an activation key JWT.",
        "operationId": "register_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterAgentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Client or validation error"
          },
          "401": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Unauthorized"
          },
          "403": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Forbidden"
          },
          "404": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Not found"
          },
          "409": {
            "description": "Activation key already used"
          },
          "500": {
            "$ref": "#/components/schemas/ApiErrorResponse",
            "summary": "Internal server error"
          },
          "503": {
            "description": "CA not initialized"
          }
        },
        "security": [
          {
            "activation_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivationKey": {
        "type": "object",
        "description": "An activation key used to bootstrap host agent configuration.",
        "required": [
          "id",
          "key_id",
          "description",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the key was created"
          },
          "description": {
            "type": "string",
            "description": "Human-readable description"
          },
          "id": {
            "type": "string",
            "description": "Internal database ID"
          },
          "key_id": {
            "type": "string",
            "description": "User-provided key identifier"
          }
        }
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "description": "An error response for an API endpoint. This is used to return errors to the\nclient in a consistent format.",
        "required": [
          "message"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "An optional error code that can be used to identify the type of error\nthat occurred."
          },
          "details": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string",
            "description": "A human-readable message describing the error that occurred."
          }
        }
      },
...
      "CreateActivationKeyRequest": {
        "type": "object",
        "description": "Request body for creating an activation key.",
        "required": [
          "key_id",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "Human-readable description"
          },
          "key_id": {
            "type": "string",
            "description": "User-provided identifier for this key"
          }
        }
      },
      "CreateActivationKeyResponse": {
        "type": "object",
        "description": "Response for activation key creation - includes the JWT token.",
        "required": [
          "key",
          "token"
        ],
        "properties": {
          "key": {
            "$ref": "#/components/schemas/ActivationKey",
            "description": "The created activation key metadata"
          },
          "token": {
            "type": "string",
            "description": "The JWT token - only returned on creation, store it securely!"
          }
        }
      },
...
      "PaginatedList_ActivationKey": {
        "type": "object",
        "description": "Parameters for paginating through a list of records. This is used by the\nvarious list endpoints to allow clients to paginate through large sets of\nrecords.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "An activation key used to bootstrap host agent configuration.",
              "required": [
                "id",
                "key_id",
                "description",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "description": "When the key was created"
                },
                "description": {
                  "type": "string",
                  "description": "Human-readable description"
                },
                "id": {
                  "type": "string",
                  "description": "Internal database ID"
                },
                "key_id": {
                  "type": "string",
                  "description": "User-provided key identifier"
                }
              }
            }
          },
          "limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "The maximum number of results to return.",
            "minimum": 0
          },
          "next_token": {
            "type": [
              "string",
              "null"
            ],
            "description": "The next page token, if any. This is acquired by requesting a paginated\nset of records and looking at the `next_token` or `prev_token` field."
          }
        }
      },
...
      "RegisterAgentRequest": {
        "type": "object",
        "required": [
          "csr_pem",
          "hostname"
        ],
        "properties": {
          "csr_pem": {
            "type": "string",
            "description": "CSR in PEM format"
          },
          "hostname": {
            "type": "string",
            "description": "Hostname of the agent"
          }
        }
      },
      "RegisterAgentResponse": {
        "type": "object",
        "required": [
          "agent_id",
          "certificate_pem",
          "ca_certificate_pem",
          "expires_at",
          "api_base_url"
        ],
        "properties": {
          "agent_id": {
            "type": "string",
            "description": "Agent UUID (ObjectId as hex string)"
          },
          "api_base_url": {
            "type": "string",
            "description": "API base URL for future requests"
          },
          "ca_certificate_pem": {
            "type": "string",
            "description": "CA certificate in PEM format"
          },
          "certificate_pem": {
            "type": "string",
            "description": "Signed certificate in PEM format"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "description": "Certificate expiration time"
          }
        }
      },
...
    }
  }
}